refactor: extract ConsoleUi and GitHelper from Program.cs
Move UI helpers (spinner, progress bar, banner, easter egg messages,
version loading, usage text) to Cli/ConsoleUi.cs and git operations
to Cli/GitHelper.cs. Program.cs reduced from 959 to 686 lines,
focusing on command routing, argument parsing, and business logic.
