+++ src/CodeIndex/Program.cs
@@ -2,6 +2,7 @@ using System.Diagnostics;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using CodeIndex.Cli;
 using CodeIndex.Database;
 using CodeIndex.Indexer;
 
@@ -11,7 +12,7 @@ using CodeIndex.Indexer;
 Console.OutputEncoding = Encoding.UTF8;
 
 // Load version from version.json / version.jsonからバージョンを読み込み
-var appVersion = LoadVersion();
+var appVersion = ConsoleUi.LoadVersion();
 
 // Exit codes / 終了コード
 // 0 = success, 1 = usage error, 2 = not found, 3 = database error
@@ -31,7 +32,7 @@ var jsonOptions = new JsonSerializerOptions
 // Route to subcommand / サブコマンドにルーティング
 if (args.Length == 0 || args[0] is "--help" or "-h")
 {
-    PrintUsage();
+    ConsoleUi.PrintUsage();
     return args.Length == 0 ? ExitUsageError : ExitSuccess;
 }
 
@@ -46,7 +47,7 @@ if (args[0] is "--version" or "-V")
 var easterEgg = args.FirstOrDefault(a => a is "--sushi" or "--coffee" or "--ramen" or "--wine" or "--beer" or "--matcha" or "--whisky");
 if (easterEgg != null && !args.Any(a => !a.StartsWith('-')))
 {
-    PrintEasterEggMessage(easterEgg);
+    ConsoleUi.PrintEasterEggMessage(easterEgg);
     return ExitSuccess;
 }
 
@@ -200,11 +201,11 @@ int RunStatus(string[] cmdArgs)
 int RunIndex(string[] indexArgs)
 {
     var (projectPath, dbPath, rebuild, verbose, jsonOutput, commits, updateFiles, easterEgg) = ParseIndexArgs(indexArgs);
-    var spinnerFrames = GetSpinnerFrames(easterEgg);
+    var spinnerFrames = ConsoleUi.GetSpinnerFrames(easterEgg);
 
     if (projectPath == null)
     {
-        PrintUsage();
+        ConsoleUi.PrintUsage();
         return ExitUsageError;
     }
 
@@ -214,7 +215,7 @@ int RunIndex(string[] indexArgs)
 
     if (!jsonOutput)
     {
-        PrintBanner();
+        ConsoleUi.PrintBanner();
         Console.WriteLine($"  Project : {Path.GetFullPath(projectPath)}");
         Console.WriteLine($"  Output  : {dbPath}");
         Console.WriteLine($"  Mode    : {mode}");
@@ -275,14 +276,14 @@ int RunUpdateMode(DbWriter writer, FileIndexer indexer, string projectRoot, stri
     {
         CancellationTokenSource? spinnerCts = null;
         if (!jsonOutput)
-            spinnerCts = StartSpinner("Resolving changed files...", spinnerFrames);
+            spinnerCts = ConsoleUi.StartSpinner("Resolving changed files...", spinnerFrames);
         foreach (var commit in commits)
         {
-            var changedFiles = GetChangedFilesFromCommit(projectRoot, commit);
+            var changedFiles = GitHelper.GetChangedFilesFromCommit(projectRoot, commit);
             foreach (var f in changedFiles)
                 targetPaths.Add(f);
         }
-        StopSpinner(spinnerCts);
+        ConsoleUi.StopSpinner(spinnerCts);
         if (!jsonOutput)
             Console.WriteLine($"  Found {targetPaths.Count} changed file(s) from git");
     }
@@ -408,9 +409,9 @@ int RunFullScan(DbWriter writer, FileIndexer indexer, string projectRoot, string
 {
     CancellationTokenSource? spinnerCts = null;
     if (!jsonOutput)
-        spinnerCts = StartSpinner("Scanning...", spinnerFrames);
+        spinnerCts = ConsoleUi.StartSpinner("Scanning...", spinnerFrames);
     var files = indexer.ScanFiles();
-    StopSpinner(spinnerCts);
+    ConsoleUi.StopSpinner(spinnerCts);
     if (!jsonOutput)
     {
         Console.WriteLine($"  Found {files.Count:N0} files");
@@ -419,15 +420,15 @@ int RunFullScan(DbWriter writer, FileIndexer indexer, string projectRoot, string
 
     CancellationTokenSource? purgeCts = null;
     if (!jsonOutput)
-        purgeCts = StartSpinner("Cleaning up stale entries...", spinnerFrames);
+        purgeCts = ConsoleUi.StartSpinner("Cleaning up stale entries...", spinnerFrames);
     var purged = writer.PurgeStaleFiles(projectRoot);
-    StopSpinner(purgeCts);
+    ConsoleUi.StopSpinner(purgeCts);
     if (purged > 0 && !jsonOutput)
         Console.WriteLine($"  Purged {purged:N0} stale files (no longer on disk)");
 
     CancellationTokenSource? indexCts = null;
     if (!jsonOutput)
-        indexCts = StartSpinner("Indexing...", spinnerFrames);
+        indexCts = ConsoleUi.StartSpinner("Indexing...", spinnerFrames);
     int processed = 0, skipped = 0, errors = 0;
     var errorList = new List<object>();
     // Stop the indexing spinner before the first progress update / 最初のプログレス更新前にスピナーを停止
@@ -437,7 +438,7 @@ int RunFullScan(DbWriter writer, FileIndexer indexer, string projectRoot, string
     {
         if (!indexSpinnerStopped)
         {
-            StopSpinner(indexCts);
+            ConsoleUi.StopSpinner(indexCts);
             indexSpinnerStopped = true;
             if (!jsonOutput) Console.WriteLine("Indexing...");
         }
@@ -452,7 +453,7 @@ int RunFullScan(DbWriter writer, FileIndexer indexer, string projectRoot, string
                 processed++;
                 if (verbose && !jsonOutput)
                     Console.WriteLine($"  [SKIP] {record.Path}");
-                if (!jsonOutput) PrintProgress(processed, files.Count);
+                if (!jsonOutput) ConsoleUi.PrintProgress(processed, files.Count);
                 continue;
             }
 
@@ -482,13 +483,13 @@ int RunFullScan(DbWriter writer, FileIndexer indexer, string projectRoot, string
         }
 
         processed++;
-        if (!jsonOutput) PrintProgress(processed, files.Count);
+        if (!jsonOutput) ConsoleUi.PrintProgress(processed, files.Count);
     }
 
     // If no files to process, stop the indexing spinner / ファイルがない場合はスピナーを停止
     if (!indexSpinnerStopped)
     {
-        StopSpinner(indexCts);
+        ConsoleUi.StopSpinner(indexCts);
         if (!jsonOutput) Console.WriteLine("Indexing...");
     }
 
@@ -682,285 +683,4 @@ static (string? projectPath, string dbPath, bool rebuild, bool verbose, bool jso
     }
 
     return (projectPath, dbPath, rebuild, verbose, json, commits, updateFiles, easterEgg);
-}
-
-// Print inline progress bar / インライン進捗バーを表示
-static void PrintProgress(int current, int total)
-{
-    // Update every 50 files or at completion / 50ファイルごと、または完了時に更新
-    if (current % 50 != 0 && current != total)
-        return;
-
-    const int barWidth = 32;
-    var pct = (double)current / total;
-    int filled = (int)Math.Round(pct * barWidth);
-    if (filled > barWidth) filled = barWidth;
-
-    var bar = new string('\u2588', filled) + new string('\u2591', barWidth - filled);
-    var line = $"  {bar} {pct * 100,5:F1}%  [{current:N0}/{total:N0}]";
-
-    if (!Console.IsOutputRedirected)
-    {
-        Console.Write($"\r{line}");
-        Console.Out.Flush();
-        if (current == total)
-            Console.WriteLine();
-    }
-    else
-    {
-        // Fallback for redirected output / リダイレクト時はフォールバック
-        Console.WriteLine(line.TrimStart());
-    }
-}
-
-// Load version from version.json / version.jsonからバージョンを読み込み
-static string LoadVersion()
-{
-    var exeDir = AppContext.BaseDirectory;
-    var path = Path.Combine(exeDir, "version.json");
-    if (!File.Exists(path))
-    {
-        // Fallback: look relative to current directory / カレントディレクトリからの相対パスでフォールバック
-        path = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "version.json");
-    }
-    if (File.Exists(path))
-    {
-        var json = File.ReadAllText(path);
-        using var doc = JsonDocument.Parse(json);
-        if (doc.RootElement.TryGetProperty("version", out var ver))
-            return ver.GetString() ?? "0.0.0";
-    }
-    return "0.0.0";
-}
-
-// Print ASCII-art banner / ASCIIアートバナーを表示
-static void PrintBanner()
-{
-    const string banner = """
-
-         ██████╗ ██████╗ ██████╗ ███████╗██╗███╗   ██╗██████╗ ███████╗██╗  ██╗
-        ██╔════╝██╔═══██╗██╔══██╗██╔════╝██║████╗  ██║██╔══██╗██╔════╝╚██╗██╔╝
-        ██║     ██║   ██║██║  ██║█████╗  ██║██╔██╗ ██║██║  ██║█████╗   ╚███╔╝
-        ██║     ██║   ██║██║  ██║██╔══╝  ██║██║╚██╗██║██║  ██║██╔══╝   ██╔██╗
-        ╚██████╗╚██████╔╝██████╔╝███████╗██║██║ ╚████║██████╔╝███████╗██╔╝ ██╗
-         ╚═════╝ ╚═════╝ ╚═════╝ ╚══════╝╚═╝╚═╝  ╚═══╝╚═════╝ ╚══════╝╚═╝  ╚═╝
-        """;
-    Console.WriteLine(banner);
-}
-
-// Get changed files from a git commit / gitコミットから変更ファイルを取得
-static List<string> GetChangedFilesFromCommit(string projectRoot, string commitId)
-{
-    // Validate commit ID to prevent argument injection (only hex + common ref chars allowed)
-    // コミットIDをバリデーションし引数インジェクションを防止（16進数+一般的な参照文字のみ許可）
-    if (!System.Text.RegularExpressions.Regex.IsMatch(commitId, @"^[a-zA-Z0-9_./^~\-]+$"))
-        throw new ArgumentException($"Invalid commit ID: {commitId}");
-
-    var psi = new ProcessStartInfo
-    {
-        FileName = "git",
-        // Use "--" to terminate options, preventing commitId from being parsed as a flag
-        // "--"でオプション終了を明示し、commitIdがフラグとして解釈されるのを防止
-        Arguments = $"diff-tree --no-commit-id -r --name-only -- {commitId}",
-        WorkingDirectory = projectRoot,
-        RedirectStandardOutput = true,
-        RedirectStandardError = true,
-        UseShellExecute = false,
-        CreateNoWindow = true,
-    };
-
-    using var process = Process.Start(psi)
-        ?? throw new InvalidOperationException("Failed to start git process / gitプロセスの起動に失敗");
-    var output = process.StandardOutput.ReadToEnd();
-    var error = process.StandardError.ReadToEnd();
-    process.WaitForExit();
-
-    if (process.ExitCode != 0)
-        throw new InvalidOperationException($"git diff-tree failed for commit {commitId}: {error.Trim()}");
-
-    return output.Split('\n', StringSplitOptions.RemoveEmptyEntries)
-        .Select(p => p.Replace('\\', '/'))
-        .ToList();
-}
-
-// Print usage information / 使い方を表示
-static void PrintUsage()
-{
-    PrintBanner();
-    Console.WriteLine("Usage: cdidx <command> [options]");
-    Console.WriteLine();
-    Console.WriteLine("Commands:");
-    Console.WriteLine("  index <projectPath>        Index a project (default if path given)");
-    Console.WriteLine("  search <query>             Full-text search across indexed chunks");
-    Console.WriteLine("  symbols [query]            Search symbols (functions, classes, imports)");
-    Console.WriteLine("  files [query]              List indexed files");
-    Console.WriteLine("  status                     Show database statistics");
-    Console.WriteLine();
-    Console.WriteLine("Index options:");
-    Console.WriteLine("  --db <path>                Database file path (default: codeindex.db)");
-    Console.WriteLine("  --rebuild                  Delete existing DB and rebuild from scratch");
-    Console.WriteLine("  --verbose                  Show per-file status ([OK  ]/[SKIP]/[DEL ]/[ERR ])");
-    Console.WriteLine("  --json                     Output results as JSON (for AI/machine use)");
-    Console.WriteLine("  --commits <id> [id ...]    Update only files changed in the specified git commits");
-    Console.WriteLine("  --files <path> [path ...]  Update only the specified files (relative or absolute)");
-    Console.WriteLine("  --help, -h                 Show this help message");
-    Console.WriteLine("  --version, -V              Show version information");
-    Console.WriteLine();
-    Console.WriteLine("Query options:");
-    Console.WriteLine("  --db <path>                Database file path (default: codeindex.db)");
-    Console.WriteLine("  --json                     Output as JSON lines (for AI/machine use)");
-    Console.WriteLine("  --limit <n>                Max results to return (default: 20)");
-    Console.WriteLine("  --lang <lang>              Filter by language");
-    Console.WriteLine("  --kind <kind>              Filter symbols by kind (function/class/import)");
-    Console.WriteLine();
-    Console.WriteLine("Examples:");
-    Console.WriteLine("  cdidx ./myproject                             Index a project");
-    Console.WriteLine("  cdidx search \"authenticate\" --db codeindex.db  Full-text search");
-    Console.WriteLine("  cdidx symbols UserService --kind class         Find class definitions");
-    Console.WriteLine("  cdidx files --lang python                      List Python files");
-    Console.WriteLine("  cdidx status --json                            DB stats as JSON");
-    Console.WriteLine();
-    Console.WriteLine("  cdidx ./myproject --commits abc123 def456      Update files from commits");
-    Console.WriteLine("  cdidx ./myproject --files src/app.cs           Update specific files");
-    Console.WriteLine();
-    Console.WriteLine("Easter eggs (themed spinner when combined with index):");
-    Console.WriteLine("  --sushi                    \U0001f363 Slicing... Shaping... Itadakimasu!");
-    Console.WriteLine("  --coffee                   \u2615 Grinding... Heating... Brewing...");
-    Console.WriteLine("  --ramen                    \U0001f35c Boiling... Steaming... Itadakimasu!");
-    Console.WriteLine("  --wine                     \U0001f377 Crushing... Aging... Sant\u00e9!");
-    Console.WriteLine("  --beer                     \U0001f37a Tapping... Pouring... Cheers!");
-    Console.WriteLine("  --matcha                   \U0001f375 Sifting... Whisking... Douzo!");
-    Console.WriteLine("  --whisky                   \U0001f943 Mashing... Distilling... Slainte!");
-    Console.WriteLine("  --random-spinner           \U0001f3b2 Pick a random theme");
-}
-
-// --- Spinner / スピナー ---
-
-// Start spinner on a background thread, returns CancellationTokenSource to stop it
-// バックグラウンドスレッドでスピナーを開始。停止用のCancellationTokenSourceを返す
-// When themed frames are used (easter egg), the frame itself contains the full display text
-// テーマフレーム使用時（イースターエッグ）はフレーム自体が表示テキストを含む
-static CancellationTokenSource? StartSpinner(string message, string[] frames)
-{
-    // Braille frames are single-char; themed frames are longer strings containing the display text
-    // ブレイルフレームは1文字、テーマフレームは表示テキストを含む長い文字列
-    bool isThemed = frames.Length > 0 && frames[0].Length > 2;
-
-    if (Console.IsOutputRedirected)
-    {
-        Console.WriteLine(message);
-        return null;
-    }
-
-    var cts = new CancellationTokenSource();
-    var ct = cts.Token;
-    Task.Run(() =>
-    {
-        int i = 0;
-        while (!ct.IsCancellationRequested)
-        {
-            var frame = frames[i % frames.Length];
-            var line = isThemed ? $"\r{frame}" : $"\r{frame} {message}";
-            Console.Write(line);
-            Console.Out.Flush();
-            i++;
-            try { Task.Delay(100, ct).Wait(ct); } catch (OperationCanceledException) { break; }
-        }
-    }, ct);
-    return cts;
-}
-
-// Stop spinner and clear the line / スピナーを停止して行をクリア
-static void StopSpinner(CancellationTokenSource? cts)
-{
-    if (cts == null) return;
-    cts.Cancel();
-    // Small delay to let the spinner task exit / スピナータスク終了のための短い待機
-    Thread.Sleep(20);
-    if (!Console.IsOutputRedirected)
-    {
-        Console.Write($"\r{new string(' ', Console.WindowWidth > 0 ? Console.WindowWidth - 1 : 80)}\r");
-        Console.Out.Flush();
-    }
-    cts.Dispose();
-}
-
-// Get spinner frames based on easter egg flag / イースターエッグフラグに基づくスピナーフレームを取得
-static string[] GetSpinnerFrames(string? easterEgg) => easterEgg switch
-{
-    "--sushi" =>
-    [
-        "\U0001f363 Slicing       ", "\U0001f363 Slicing.      ", "\U0001f363 Slicing..     ", "\U0001f363 Slicing...    ",
-        "\U0001f363 Shaping       ", "\U0001f363 Shaping.      ", "\U0001f363 Shaping..     ", "\U0001f363 Shaping...    ",
-        "\U0001f363 Pressing      ", "\U0001f363 Pressing.     ", "\U0001f363 Pressing..    ", "\U0001f363 Pressing...   ",
-        "\U0001f363 Itadakimasu!  ",
-    ],
-    "--coffee" =>
-    [
-        "\u2615 Grinding      ", "\u2615 Grinding.     ", "\u2615 Grinding..    ", "\u2615 Grinding...   ",
-        "\u2615 Heating       ", "\u2615 Heating.      ", "\u2615 Heating..     ", "\u2615 Heating...    ",
-        "\u2615 Brewing       ", "\u2615 Brewing.      ", "\u2615 Brewing..     ", "\u2615 Brewing...    ",
-    ],
-    "--ramen" =>
-    [
-        "\U0001f35c Boiling       ", "\U0001f35c Boiling.      ", "\U0001f35c Boiling..     ", "\U0001f35c Boiling...    ",
-        "\U0001f35c Steaming      ", "\U0001f35c Steaming.     ", "\U0001f35c Steaming..    ", "\U0001f35c Steaming...   ",
-        "\U0001f35c Slurping      ", "\U0001f35c Slurping.     ", "\U0001f35c Slurping..    ", "\U0001f35c Slurping...   ",
-        "\U0001f35c Itadakimasu!  ",
-    ],
-    "--wine" =>
-    [
-        "\U0001f377 Crushing      ", "\U0001f377 Crushing.     ", "\U0001f377 Crushing..    ", "\U0001f377 Crushing...   ",
-        "\U0001f377 Aging         ", "\U0001f377 Aging.        ", "\U0001f377 Aging..       ", "\U0001f377 Aging...      ",
-        "\U0001f377 Pouring       ", "\U0001f377 Pouring.      ", "\U0001f377 Pouring..     ", "\U0001f377 Pouring...    ",
-        "\U0001f377 Sant\u00e9!        ",
-    ],
-    "--beer" =>
-    [
-        "\U0001f37a Tapping       ", "\U0001f37a Tapping.      ", "\U0001f37a Tapping..     ", "\U0001f37a Tapping...    ",
-        "\U0001f37a Pouring       ", "\U0001f37a Pouring.      ", "\U0001f37a Pouring..     ", "\U0001f37a Pouring...    ",
-        "\U0001f37a Foaming       ", "\U0001f37a Foaming.      ", "\U0001f37a Foaming..     ", "\U0001f37a Foaming...    ",
-        "\U0001f37a Cheers!       ",
-    ],
-    "--matcha" =>
-    [
-        "\U0001f375 Sifting       ", "\U0001f375 Sifting.      ", "\U0001f375 Sifting..     ", "\U0001f375 Sifting...    ",
-        "\U0001f375 Pouring       ", "\U0001f375 Pouring.      ", "\U0001f375 Pouring..     ", "\U0001f375 Pouring...    ",
-        "\U0001f375 Whisking      ", "\U0001f375 Whisking.     ", "\U0001f375 Whisking..    ", "\U0001f375 Whisking...   ",
-        "\U0001f375 Douzo!        ",
-    ],
-    "--whisky" =>
-    [
-        "\U0001f943 Mashing       ", "\U0001f943 Mashing.      ", "\U0001f943 Mashing..     ", "\U0001f943 Mashing...    ",
-        "\U0001f943 Distilling    ", "\U0001f943 Distilling.   ", "\U0001f943 Distilling..  ", "\U0001f943 Distilling... ",
-        "\U0001f943 Aging         ", "\U0001f943 Aging.        ", "\U0001f943 Aging..       ", "\U0001f943 Aging...      ",
-        "\U0001f943 Slainte!      ",
-    ],
-    // Default: Braille spinner / デフォルト: ブレイルスピナー
-    _ => ["\u2807", "\u280b", "\u280f", "\u2838", "\u280f", "\u2839"],
-};
-
-// Print easter egg message (standalone mode) / イースターエッグメッセージを表示（単体実行時）
-static void PrintEasterEggMessage(string flag)
-{
-    var (en, ja) = flag switch
-    {
-        "--sushi"  => ("\U0001f363 Indexing is like making sushi \u2014 patience yields perfection.",
-                       "   \u30a4\u30f3\u30c7\u30c3\u30af\u30b9\u306f\u5bff\u53f8\u4f5c\u308a\u306e\u3088\u3046\u306b \u2014 \u5fcd\u8010\u304c\u5b8c\u74a7\u3092\u751f\u3080\u3002"),
-        "--coffee" => ("\u2615 Leave the indexing to me and go grab a coffee!",
-                       "   \u30a4\u30f3\u30c7\u30c3\u30af\u30b9\u306f\u4efb\u305b\u3066\u3001\u30b3\u30fc\u30d2\u30fc\u3067\u3082\u98f2\u3093\u3067\u304d\u3066\uff01"),
-        "--ramen"  => ("\U0001f35c Indexing in progress... perfect time for a bowl of ramen!",
-                       "   \u30a4\u30f3\u30c7\u30c3\u30af\u30b9\u4e2d\u2026\u30e9\u30fc\u30e1\u30f3\u4e00\u676f\u3044\u304b\u304c\uff1f"),
-        "--wine"   => ("\U0001f377 Crushing... Aging... Pouring... Sant\u00e9!",
-                       "   \u30a4\u30f3\u30c7\u30c3\u30af\u30b9\u306f\u30ef\u30a4\u30f3\u306e\u3088\u3046\u306b\u2014\u719f\u6210\u3092\u5f85\u3064\u4fa1\u5024\u304c\u3042\u308b\u3002"),
-        "--beer"   => ("\U0001f37a Tapping... Pouring... Foaming... Cheers!",
-                       "   \u30a4\u30f3\u30c7\u30c3\u30af\u30b9\u5b8c\u4e86\u307e\u3067\u3001\u4e7e\u676f\uff01"),
-        "--matcha" => ("\U0001f375 Sifting... Pouring... Whisking... \u3069\u3046\u305e\uff01",
-                       "   \u4e00\u670d\u306e\u62b9\u8336\u3067\u3082\u3044\u304b\u304c\u3067\u3059\u304b\uff1f"),
-        "--whisky" => ("\U0001f943 Mashing... Distilling... Aging... Slainte!",
-                       "   \u30a4\u30f3\u30c7\u30c3\u30af\u30b9\u306f\u30a6\u30a4\u30b9\u30ad\u30fc\u306e\u3088\u3046\u306b\u2014\u719f\u6210\u304c\u5927\u4e8b\u3002"),
-        _ => ("", ""),
-    };
-    Console.WriteLine(en);
-    Console.WriteLine(ja);
 }
